mock platform, tests refactored
