+++ lib/platform/index.js
@@ -8,8 +8,6 @@ if (target === 'android' || target === 'ios') {
 // temp until env works!
 } else if (agent.platform === 'ios' || agent.platform === 'android') {
   module.exports = require('./native')
-} else if (agent.browser) {
-  module.exports = require('./native')
 } else {
-  module.exports = { _platform: {} }
+  module.exports = require('./mock')
 }
