+++ lib/platform/mock.js
